[FIRRTL][IMDCE] Delete dead ref ops.
When deleting ports, IMDCE introduces temporary wires for the ports.
These are invalid for ref type, which is bad.

Definitely don't risk leaving such a wire around, but that is hard
to introduce: must have an output ref port that's alive but not used,
which is currently illegal for ref's (but check anyway to be sure).