[FIRRTL][SV] Use DefaultValuedOptionalAttr, restore previous behavior.
Changes to DefaultValuedAttr made these no longer optional,
but omitting them when not specified is nice for readability.

For operations that we custom parse/print this may matter less,
but for a few this cleans up / preserves the output.

Only DefaultValuedAttr remaining is the "resetStyle" for sv.alwaysff.
The generated code doesn't compile when this is made optional,
but also it makes sense to ask for the resetStyle (default: "NoReset")
when resetEdge is missing.

The only test reflecting this behavior either way (for our ops)
is "expand-whens", update accordingly.

sv.verbatim will omit the symbols array when not specified,
but round-tripping it with `{ symbols = [] }` still includes it.
This may be a difference in null attribute vs empty array (?).