[FIRRTL] Improve IMCP for non-hw; support materializing refs of constants.
Fixes #6076.

Fixes leaving dead property wires around if they're constant-prop'd through.

Tweak export-ref test to demonstrate somewhat that this now enables
a probe of a constant in a module to be hoisted to a probe of the constant
at the boundary, allow for more cleanup -- in particular prevents modules
with probed constants being preserved just for that reason.
(The test adds a dontTouch to avoid this to check the exported macro)